+++ src/index.js
@@ -1,5 +1,5 @@
-const matcher = require('./matcher')
-const serializer = require('./serializer')
+const matcher = require('./matchers/toMatchStyledComponentsSnapshot')
+const serializer = require('./serializers/styleSheetSerializer')
 
 module.exports = {
   matcher,
