+++ src/matchers/toHaveStyleRule.js
