+++ src/matchers/toMatchStyledComponentsSnapshot.js
