--- src/native/index.js
@@ -1,5 +1,0 @@
-const toHaveStyleRule = require('./toHaveStyleRule')
-
-module.exports = {
-  toHaveStyleRule,
-}
